+++ HISTORY.rst
@@ -478,7 +478,10 @@ Bug fixes
 - The "reload" method of Config overwrites the values and thus gets rid of the "context" field which is
   only set in the constructor of the class.
 - Fixed all the test cases, such that they *should* all pass now.
-- AcquireSingleFrameTest fails because a call to plt.subplots gets the wrong parameters.
+- "tests.frame.AcquireSingleFrame" fails because a call to plt.subplots gets the wrong parameters.
+- "tests.frame.AcquireSingleFrame" fails sometimes if the pixel values are too homogeneous and the difference of the
+  0.1 and 0.9 quantile values comes down to zero. The procedure to increase contrast then causes a division by
+  zero exception
 
 
 TODO